refactor: Update styles.scss to modify form-floating label color and position
The styles.scss file has been updated to modify the color and position of the label in the form-floating class. The color has been changed to #6c757d and the left position has been set to 10px. This change improves the visual appearance and consistency of the form elements.

Note: This commit message follows the established convention of starting with a type (refactor) and providing a concise and descriptive summary of the changes made.