refactor: Update package.json and package-lock.json to add @ngx-translate/core dependency
This commit adds the "@ngx-translate/core" dependency to the package.json and package-lock.json files. The added dependency is at version "^15.0.0". This change is necessary to support the implementation of language translation features in the application.

Note: This commit message follows the established convention of starting with a type (refactor) and providing a concise and descriptive summary of the changes made.